chore(exporter-otlp-*-http): clean up tests
